backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
python-decouple==3.8
openai==1.3.0

# Testing dependencies
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
faker==22.0.0
